backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
pydantic[email]
python-jose
passlib[bcrypt]
python-multipart
boto3
celery
redis

openai-whisper
torch
sentence-transformers

pgvector
alembic
sqlalchemy-utils

requests
httpx

python-dotenv
google-generativeai
groq
pydub

# Phase 1: PDF Export & Multilingual
reportlab
google-cloud-translate
PyPDF2

# Phase 2: Document Processing & Chatbot
python-docx

# Phase 3: Live Meeting WebSocket
websockets
python-socketio
python-engineio

# Phase 4: Speaker Diarization
pyannote.audio
